+++ operator/pkg/eventfilter_test.go
@@ -5,6 +5,8 @@ package pkg
 
 import (
 	"context"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"strings"
 
 	kartav1alpha1 "github.com/run-ai/karta/pkg/api/runai/v1alpha1"
 
@@ -108,3 +110,44 @@ var _ = Describe("Reconciler.MapCRDToKartaEvent", func() {
 		Expect(r.MapCRDToKartaEvent(ctx, crd)).To(BeEmpty())
 	})
 })
+
+// newKarta builds a Karta with the given name and optional root GVK. The root
+// component has no StatusDefinition so it will fail spec validation — use
+// newValidKarta when Validated=True is expected.
+func newKarta(name string, gvk *schema.GroupVersionKind) *kartav1alpha1.Karta {
+	k := &kartav1alpha1.Karta{
+		ObjectMeta: metav1.ObjectMeta{Name: name},
+	}
+	if gvk != nil {
+		k.Spec.StructureDefinition.RootComponent.Name = name + "-root"
+		k.Spec.StructureDefinition.RootComponent.Kind = &kartav1alpha1.GroupVersionKind{
+			Group:   gvk.Group,
+			Version: gvk.Version,
+			Kind:    gvk.Kind,
+		}
+	}
+	return k
+}
+
+// newCRD builds a CustomResourceDefinition serving the given versions for
+// the supplied group/kind. The first version is marked as storage.
+func newCRD(name, group, kind string, versions ...string) *apiextensionsv1.CustomResourceDefinition {
+	v := make([]apiextensionsv1.CustomResourceDefinitionVersion, 0, len(versions))
+	for _, ver := range versions {
+		v = append(v, apiextensionsv1.CustomResourceDefinitionVersion{
+			Name: ver, Served: true, Storage: ver == versions[0],
+		})
+	}
+	return &apiextensionsv1.CustomResourceDefinition{
+		ObjectMeta: metav1.ObjectMeta{Name: name},
+		Spec: apiextensionsv1.CustomResourceDefinitionSpec{
+			Group: group,
+			Names: apiextensionsv1.CustomResourceDefinitionNames{
+				Kind:   kind,
+				Plural: strings.ToLower(kind) + "s",
+			},
+			Scope:    apiextensionsv1.ClusterScoped,
+			Versions: v,
+		},
+	}
+}
