+++ operator/pkg/suite_test.go
@@ -4,7 +4,6 @@
 package pkg
 
 import (
-	"strings"
 	"testing"
 
 	kartav1alpha1 "github.com/run-ai/karta/pkg/api/runai/v1alpha1"
@@ -12,9 +11,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -38,56 +35,4 @@ func buildScheme() *runtime.Scheme {
 // the test explicitly reads from the recorder channel.
 func newReconciler(k8s client.WithWatch) *Reconciler {
 	return NewReconciler(k8s, record.NewFakeRecorder(64))
-}
-
-// newKarta builds a Karta with the given name and optional root GVK. The root
-// component has no StatusDefinition so it will fail spec validation — use
-// newValidKarta when Validated=True is expected.
-func newKarta(name string, gvk *schema.GroupVersionKind) *kartav1alpha1.Karta {
-	k := &kartav1alpha1.Karta{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-	}
-	if gvk != nil {
-		k.Spec.StructureDefinition.RootComponent.Name = name + "-root"
-		k.Spec.StructureDefinition.RootComponent.Kind = &kartav1alpha1.GroupVersionKind{
-			Group:   gvk.Group,
-			Version: gvk.Version,
-			Kind:    gvk.Kind,
-		}
-	}
-	return k
-}
-
-// findCondition returns the condition with the given type, failing the test
-// when not found.
-func findCondition(conds []metav1.Condition, t kartav1alpha1.ConditionType) metav1.Condition {
-	GinkgoHelper()
-	c, found := findConditionOpt(conds, t)
-	if !found {
-		Fail("condition not found: " + string(t))
-	}
-	return c
-}
-
-// newCRD builds a CustomResourceDefinition serving the given versions for
-// the supplied group/kind. The first version is marked as storage.
-func newCRD(name, group, kind string, versions ...string) *apiextensionsv1.CustomResourceDefinition {
-	v := make([]apiextensionsv1.CustomResourceDefinitionVersion, 0, len(versions))
-	for _, ver := range versions {
-		v = append(v, apiextensionsv1.CustomResourceDefinitionVersion{
-			Name: ver, Served: true, Storage: ver == versions[0],
-		})
-	}
-	return &apiextensionsv1.CustomResourceDefinition{
-		ObjectMeta: metav1.ObjectMeta{Name: name},
-		Spec: apiextensionsv1.CustomResourceDefinitionSpec{
-			Group: group,
-			Names: apiextensionsv1.CustomResourceDefinitionNames{
-				Kind:   kind,
-				Plural: strings.ToLower(kind) + "s",
-			},
-			Scope:    apiextensionsv1.ClusterScoped,
-			Versions: v,
-		},
-	}
 }
